requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.7.4
python-multipart==0.0.9
Pillow==10.4.0
paddleocr==2.7.0.3
opencv-python-headless==4.10.0.84
numpy==1.26.4
qrcode==7.4.2
psutil==5.9.8

